Remove compiled caches
