--- srsRAN_5G/srsRAN_5G/simulation_results/epc_stdout.log
@@ -1,7 +1,0 @@
-
-Built in Release mode using commit ec29b0c1f on branch master.
-
-
----  Software Radio Systems EPC  ---
-
-unrecognised option '--mme.log.filename'
